% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FizzBuzz.R
\name{fizz_buzz}
\alias{fizz_buzz}
\title{Fizz Buzz Function.}
\usage{
fizz_buzz(x)
}
\arguments{
\item{x}{A vector}
}
\value{
A vector that is built from \code{x}
}
\description{
Because this is a very simple function, my explanation is short. These
paragraphs should explain everything you need to know.
}
