% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Flagstaff_Weather.R
\docType{data}
\name{Flagstaff_Weather}
\alias{Flagstaff_Weather}
\title{A short title}
\format{
A data frame with XXX observations with ZZ columns.
\describe{
   \item{DATE}{It is the day this data was recorded}
   \item{Minimum Temperature}{The recorded lowest temperature it reached that day}
   \item{TMAX}{The recorded highest temperature it reached that day}
   \item{PRCP}{The percipitation recorded that day}
   \item{SNOW}{The amount of snow that fell that day}
}
}
\usage{
Flagstaff_Weather
}
\description{
A longer paragraph that describes the context of the dataset and
discusses important aspects that will be necessary for somebody first seeing
the data to know about. Any text in these initial paragraphs will be
in the description section of the documentation file.
}
\keyword{datasets}
